+++ chat/norris/commands.go
