+++ chat/norris/main.go
