+++ chat/openai/commands.go
