+++ game/coinflip/commands.go
