+++ game/coinflip/main.go
